Systray fixed (tobe tested on W8)
The systray should be only be shown when the OSLevel is < Windows 8.
The fix won't work if the taskbarIcon's constructor throws the exception tough...
In this case we'll have to rely on a way dirtier way : creating the taskbar icon in codebehind only when running on the right OS.
If none of the above work, we'll have to find another library that handles WPF systray applications.