The editor now updates the RequirementLayer according to the KeyCommands used in its keys. (see below for more information)
The editor uses the ProtocolEditorsProvider to fetch the type of the service that is supposed to handle the keycommands found in the keys.
So we need to add the remaining editors (Mode, MonitorOnce, DynCommands) to handle the whole set of commands.
The problem is that these protocols are not following the protocol format (which is "protocol:parameter". if we change the way they work right now, the commands of the currently used CiviKey won't be handled anymore. We'll to figure out a way to migrate them.