Implemented the pause feature (see below)
Caveat : if the user clicks on another key while pausing, all commands tailing the pause (+ the ones linked to the aforementioned key) will be executed without any additional pause.
Which means that the time needed by the user to click on one key and then the other will be "forgotten".